Storing vendors in diagram service to prevent hitting the API every time
